// Frontend/FESCFrontend/src/app/models/entities/airlines/airlines.ts
export interface Airline {
  airlineId: number;
  name: string;
  description: string;
  email: string;
  phone: string;
  address: string;
  countryId: number;
  cityId: number;
  zip: string;
  createdAt: Date;
  updatedAt: Date;
  createdByUserId: number;
  updatedByUserId: number;
  statusId: number;
  hotelId: number;
}